Setup javascript test environment
This configures the javascript test environment. It uses mocha-webpack
and enzyme with a modified version of the Rails webpack test
configuration.

The main modification is to use `webpack-node-externals` to speed up the
webpack compilation that happens at the start of each test run by not
including node modules.

It also adds a test helper that configures enzyme to use the correct
react adapter and sets up jsdom so that full enzyme rendering will work
correctly.